requirements.txt
discord.py
requests
python-dotenv
firebase-admin